+++ src/App.jsx
@@ -1,68 +1,19 @@
-import { useState } from 'react';
-import { CORE_CONCEPTS } from './data.js';
 import Header from './components/Header/Header.jsx';
-import CoreConcept from './components/CoreConcept.jsx'
-import TabButton from './components/TabButton.jsx';
-import { EXAMPLES } from './data.js'
-
+import CoreConcepts from './components/CoreConcepts.jsx';
+import Examples from './components/Examples.jsx';
 
 function App() {
-  const [selectedTopic, setSelectedTopic] = useState("components");
-
-  function handleSelect(selectedButton) {
-    setSelectedTopic(selectedButton);
-    // console.log(selectedTopic);
-  }
-
-  let tabContent = <p>Please Select a Topic.</p>;
-  if (selectedTopic) {
-    tabContent = (<div id="tab-content">
-      <h3>{EXAMPLES[selectedTopic].title}</h3>
-      <p>{EXAMPLES[selectedTopic].descripion}</p>
-      <pre>
-        <code>
-          {EXAMPLES[selectedTopic].code}
-        </code>
-      </pre>
-    </div>);
-
-  }
-
-
   return (
-    <div>
+    <>
       <Header />
       <main>
         <h2>Time to get started!</h2>
-        <section id="core-concepts">
-          <h2>Core Concepts</h2>
-          <ul>
-            {/* <CoreConcept
-              title={CORE_CONCEPTS[0].title}
-              description={CORE_CONCEPTS[0].description}
-              img={CORE_CONCEPTS[0].image}
-            /> */}
-            {/* <CoreConcept {...CORE_CONCEPTS[0]} />
-            <CoreConcept {...CORE_CONCEPTS[1]} />
-            <CoreConcept {...CORE_CONCEPTS[2]} />
-            <CoreConcept {...CORE_CONCEPTS[3]} /> */}
-            {CORE_CONCEPTS.map((conceptItem) => (
-              <CoreConcept key={conceptItem.title} {...conceptItem} /> //special propt use to remove the warning can be use explicitly 
-            ))}
-          </ul>
-        </section>
-        <section id="examples">
-          <h2>Examples</h2>
-          <menu>
-            <TabButton isSelected={selectedTopic === 'components'} onSelect={() => handleSelect('components')}>Components</TabButton>
-            <TabButton isSelected={selectedTopic === 'jsx'} onSelect={() => handleSelect('jsx')}>JSX</TabButton>
-            <TabButton isSelected={selectedTopic === 'props'} onSelect={() => handleSelect('props')}>Props</TabButton>
-            <TabButton isSelected={selectedTopic === 'state'} onSelect={() => handleSelect('state')}>State</TabButton>
-          </menu>
-          {tabContent}
-        </section>
+        <CoreConcepts/>
+        <Examples />
+       
       </main>
-    </div>
+  </>
+    
   );
 }
