+++ src/components/Examples.jsx
@@ -1,0 +1,48 @@
+import TabButton from './TabButton';
+import Section from './Section.jsx';
+import { useState } from 'react';
+import { EXAMPLES } from '../data.js'
+import Tabs from './Tabs.jsx';
+
+export default function Examples() {
+
+    const [selectedTopic, setSelectedTopic] = useState("components");
+
+    function handleSelect(selectedButton) {
+        setSelectedTopic(selectedButton);
+        // console.log(selectedTopic);
+    }
+
+    let tabContent = <p>Please Select a Topic.</p>;
+    if (selectedTopic) {
+        tabContent = (<div id="tab-content">
+            <h3>{EXAMPLES[selectedTopic].title}</h3>
+            <p>{EXAMPLES[selectedTopic].description}</p>
+            <pre>
+                <code>
+                    {EXAMPLES[selectedTopic].code}
+                </code>
+            </pre>
+        </div>);
+
+    }
+
+    return (
+        <Section title="Examples" id="examples">
+            <Tabs
+                buttonContainer="menu"
+                buttons={
+                    <>
+                        <TabButton isSelected={selectedTopic === 'components'} onClick={() => handleSelect('components')}>Components</TabButton>
+                        <TabButton isSelected={selectedTopic === 'jsx'} onClick={() => handleSelect('jsx')}>JSX</TabButton>
+                        <TabButton isSelected={selectedTopic === 'props'} onClick={() => handleSelect('props')}>Props</TabButton>
+                        <TabButton isSelected={selectedTopic === 'state'} onClick={() => handleSelect('state')}>State</TabButton>
+                    </>
+                }
+            >
+                {tabContent}
+            </Tabs>
+
+        </Section>
+    );
+}
