+++ src/components/TabButton.jsx
@@ -1,6 +1,3 @@
-export default function TabButton({ children , onSelect, isSelected }) {
-
-
-
-    return <li><button className={isSelected?'active':''} onClick={onSelect}>{children}</button></li>;
+export default function TabButton({ children , isSelected , ...props}) {
+    return <li><button className={isSelected?'active':''}{...props}>{children}</button></li>;
 }
