+++ src/components/Tabs.jsx
@@ -1,0 +1,10 @@
+export default function Tabs({children,buttons,buttonContainer}){
+
+    const ButtonContainer=buttonContainer;
+    return<>
+    <ButtonContainer>
+        {buttons}
+    </ButtonContainer>
+    {children}
+    </>
+}
